fix: 에러 해결을 위해 generateMetadata page로 이동, page 컴포넌트 서버컴포넌트로 만들기
