+++ apps/admin/app/studies/[studyId]/layout.tsx
@@ -1,16 +1,3 @@
-import { studyApi } from "apis/study/studyApi";
-
-export const generateMetadata = async ({
-  params: { studyId },
-}: {
-  params: { studyId: string };
-}) => {
-  const study = await studyApi.getStudyBasicInfo(+studyId);
-  return {
-    title: study ? study.title : "스터디",
-  };
-};
-
 const StudyLayout = ({
   children,
   modal,
