+++ apps/admin/app/studies/[studyId]/page.tsx
@@ -1,5 +1,6 @@
 import { Flex } from "@styled-system/jsx";
 import { Space } from "@wow-class/ui";
+import { studyApi } from "apis/study/studyApi";
 import Divider from "wowds-ui/Divider";
 
 import StudyAnnouncement from "./_components/announcement/StudyAnnouncement";
@@ -7,6 +8,17 @@ import AssignmentList from "./_components/assignment/AssignmentList";
 import CheckAttendanceNumber from "./_components/attendance/CheckAttendanceNumber";
 import CurriculumList from "./_components/curriculum/CurriculumList";
 import Header from "./_components/header/Header";
+
+export const generateMetadata = async ({
+  params: { studyId },
+}: {
+  params: { studyId: string };
+}) => {
+  const study = await studyApi.getStudyBasicInfo(+studyId);
+  return {
+    title: study ? study.title : "스터디",
+  };
+};
 
 const StudyPage = ({ params }: { params: { studyId: string } }) => {
   const { studyId } = params;
