+++ apps/admin/app/studies/detail-info/[studyId]/layout.tsx
@@ -1,18 +1,3 @@
-import { studyApi } from "apis/study/studyApi";
-
-export const generateMetadata = async ({
-  params: { studyId },
-}: {
-  params: { studyId: string };
-}) => {
-  const study = await studyApi.getStudyBasicInfo(+studyId);
-  return {
-    title: study
-      ? `${study.title} 스터디 상세 작성하기`
-      : "스터디 상세 작성하기",
-  };
-};
-
 const StudyDetailInfoLayout = ({
   children,
   modal,
